// apps/frontend/common/types.ts
export type TokenType = {
  _id: string;
  name: string;
  email: string;
  image: string;
  role: string;
  phone: {
    number: string;
    verified: boolean;
    otp: string;
  };
};